--- core/network-kmp/src/androidMain/kotlin/org/michaelbel/movies/network/retrofit/ktx/RetrofitKtx.kt
@@ -1,5 +1,0 @@
-package org.michaelbel.movies.network.retrofit.ktx
-
-import retrofit2.Retrofit
-
-internal inline fun <reified T> Retrofit.createService(): T = create(T::class.java)
