+++ core/network-kmp/src/commonMain/kotlin/org/michaelbel/movies/network/MovieNetworkService.kt
@@ -1,5 +1,33 @@
-@file:Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
-
 package org.michaelbel.movies.network
 
-expect class MovieNetworkService
+import org.michaelbel.movies.network.ktor.KtorMovieService
+import org.michaelbel.movies.network.model.ImagesResponse
+import org.michaelbel.movies.network.model.Movie
+import org.michaelbel.movies.network.model.MovieResponse
+import org.michaelbel.movies.network.model.Result
+
+class MovieNetworkService internal constructor(
+    private val ktorMovieService: KtorMovieService
+) {
+
+    suspend fun movies(
+        list: String,
+        language: String,
+        page: Int
+    ): Result<MovieResponse> {
+        return ktorMovieService.movies(list, language, page)
+    }
+
+    suspend fun movie(
+        movieId: Int,
+        language: String
+    ): Movie {
+        return ktorMovieService.movie(movieId, language)
+    }
+
+    suspend fun images(
+        movieId: Int
+    ): ImagesResponse {
+        return ktorMovieService.images(movieId)
+    }
+}
