+++ core/network-kmp/src/commonMain/kotlin/org/michaelbel/movies/network/SearchNetworkService.kt
@@ -1,5 +1,18 @@
-@file:Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
-
 package org.michaelbel.movies.network
 
-expect class SearchNetworkService
+import org.michaelbel.movies.network.ktor.KtorSearchService
+import org.michaelbel.movies.network.model.MovieResponse
+import org.michaelbel.movies.network.model.Result
+
+class SearchNetworkService internal constructor(
+    private val ktorSearchService: KtorSearchService
+) {
+
+    suspend fun searchMovies(
+        query: String,
+        language: String,
+        page: Int
+    ): Result<MovieResponse> {
+        return ktorSearchService.searchMovies(query, language, page)
+    }
+}
